Simplify three-agent digest proof tests and config wiring
The digest implementation added fixture hashing and new stale-artifact coverage. This refactor keeps the behavior unchanged while splitting public command config and test support into smaller files that satisfy the repo line and function budgets.\n\nConstraint: AGENTS.md requires a mandatory refactor phase after green implementation.\nConstraint: Public e2e harness API still needs missing-docs coverage under strict warning gates.\nRejected: Suppress missing_docs or dead_code globally | that would weaken the quality gates instead of cleaning the touched surfaces.\nConfidence: high\nScope-risk: narrow\nDirective: Keep stale digest regressions in the focused view digest contract instead of folding them back into broader disclosure tests.\nTested: cargo fmt --check\nTested: CARGO_TARGET_DIR=target/mvp-demo-proof CARGO_BUILD_JOBS=1 CARGO_INCREMENTAL=0 cargo clippy --workspace --all-targets --all-features -- -D warnings\nTested: CARGO_TARGET_DIR=target/mvp-demo-proof CARGO_BUILD_JOBS=1 CARGO_INCREMENTAL=0 cargo test -p kamn-e2e-harness --test mvp_demo_three_agent_view_artifact_contract --test mvp_demo_three_agent_view_digest_contract --test mvp_demo_three_agent_transcript_contract --test mvp_demo_command_contract -- --nocapture\nNot-tested: make check, local demo, and live devnet demo pending integration phase.\nRelated: #7068\n\nCo-authored-by: OmX <omx@oh-my-codex.dev>
