+++ crates/kamn-e2e-harness/src/mvp_demo/mod.rs
@@ -6,6 +6,7 @@ mod agent_harness_actor_rehearsal;
 mod agent_harness_json;
 mod agent_harness_three_agent;
 mod artifact_digest;
+mod command_config;
 mod devnet_settlement;
 mod devnet_settlement_build;
 mod devnet_settlement_json;
@@ -34,14 +35,12 @@ mod three_agent_views;
 mod verify;
 mod verify_support;
 
+pub use command_config::{MvpDemoCommandConfig, VerifyMvpDemoCommandConfig};
 pub use report::{
     CLAIM_LABEL_DEVNET_BACKED, CLAIM_LABEL_DRY_RUN, CLAIM_LABEL_LOCAL_ONLY,
     CLAIM_LABEL_PLACEHOLDER, CLAIM_LABEL_REAL, CLAIM_LABEL_ROADMAP, MVP_DEMO_REPORT_SCHEMA_VERSION,
 };
-pub use runner::{
-    execute_mvp_demo_contract, execute_verify_mvp_demo_contract, MvpDemoCommandConfig,
-    VerifyMvpDemoCommandConfig,
-};
+pub use runner::{execute_mvp_demo_contract, execute_verify_mvp_demo_contract};
 pub use verify::verify_mvp_demo_report_json;
 
 pub(crate) const DEFAULT_MVP_DEMO_OUTPUT_ROOT: &str = ".kamn/demo";
